deduplicate locale ids
for 2+ locales with the same name, we append the country code to the id for disambiguation, e.g., sydney_ca for Sydney, Canada. identified the dupes using the new schematron. the new files don't all pass the relaxng schema, but the remaining schema issues have to do with predecessor/successor fields, and can be addressed in another pass. similar changes will need to be applied to the "posts" collection.